+++ src/deepskylog/AstronomyLibrary/Targets/Target.php
@@ -65,6 +65,112 @@ class Target
     // The magnitude of the object
     private ?float $_magnitude = null;
 
+    // Needed for the calculationn of the contrast reserve
+    private $_LTCSize   = 24;
+    private $_angleSize = 7;
+    private $_angle     = [
+        -0.2255, 0.5563, 0.9859, 1.260,
+        1.742, 2.083, 2.556,
+    ];
+    private $_LTC = [
+        [
+            4, -0.3769, -1.8064, -2.3368, -2.4601,
+            -2.5469, -2.5610, -2.5660,
+        ],
+        [
+            5, -0.3315, -1.7747, -2.3337, -2.4608,
+            -2.5465, -2.5607, -2.5658,
+        ],
+        [
+            6, -0.2682, -1.7345, -2.3310, -2.4605,
+            -2.5467, -2.5608, -2.5658,
+        ],
+        [
+            7, -0.1982, -1.6851, -2.3140, -2.4572,
+            -2.5481, -2.5615, -2.5665,
+        ],
+        [
+            8, -0.1238, -1.6252, -2.2791, -2.4462,
+            -2.5463, -2.5597, -2.5646,
+        ],
+        [
+            9, -0.0424, -1.5529, -2.2297, -2.4214,
+            -2.5343, -2.5501, -2.5552,
+        ],
+        [
+            10, 0.0498, -1.4655, -2.1659, -2.3763,
+            -2.5047, -2.5269, -2.5333,
+        ],
+        [
+            11, 0.1596, -1.3581, -2.0810, -2.3036,
+            -2.4499, -2.4823, -2.4937,
+        ],
+        [
+            12, 0.2934, -1.2256, -1.9674, -2.1965,
+            -2.3631, -2.4092, -2.4318,
+        ],
+        [
+            13, 0.4557, -1.0673, -1.8186, -2.0531,
+            -2.2445, -2.3083, -2.3491,
+        ],
+        [
+            14, 0.6500, -0.8841, -1.6292, -1.8741,
+            -2.0989, -2.1848, -2.2505,
+        ],
+        [
+            15, 0.8808, -0.6687, -1.3967, -1.6611,
+            -1.9284, -2.0411, -2.1375,
+        ],
+        [
+            16, 1.1558, -0.3952, -1.1264, -1.4176,
+            -1.7300, -1.8727, -2.0034,
+        ],
+        [
+            17, 1.4822, -0.0419, -0.8243, -1.1475,
+            -1.5021, -1.6768, -1.8420,
+        ],
+        [
+            18, 1.8559, 0.3458, -0.4924, -0.8561,
+            -1.2661, -1.4721, -1.6624,
+        ],
+        [
+            19, 2.2669, 0.6960, -0.1315, -0.5510,
+            -1.0562, -1.2892, -1.4827,
+        ],
+        [
+            20, 2.6760, 1.0880, 0.2060, -0.3210,
+            -0.8800, -1.1370, -1.3620,
+        ],
+        [
+            21, 2.7766, 1.2065, 0.3467, -0.1377,
+            -0.7361, -0.9964, -1.2439,
+        ],
+        [
+            22, 2.9304, 1.3821, 0.5353, 0.0328,
+            -0.5605, -0.8606, -1.1187,
+        ],
+        [
+            23, 3.1634, 1.6107, 0.7708, 0.2531,
+            -0.3895, -0.7030, -0.9681,
+        ],
+        [
+            24, 3.4643, 1.9034, 1.0338, 0.4943,
+            -0.2033, -0.5259, -0.8288,
+        ],
+        [
+            25, 3.8211, 2.2564, 1.3265, 0.7605,
+            0.0172, -0.2992, -0.6394,
+        ],
+        [
+            26, 4.2210, 2.6320, 1.6990, 1.1320,
+            0.2860, -0.0510, -0.4080,
+        ],
+        [
+            27, 4.6100, 3.0660, 2.1320, 1.5850,
+            0.6520, 0.2410, -0.1210,
+        ],
+    ];
+
     /**
      * The constructor.
      */
@@ -1146,8 +1252,10 @@ class Target
 
     /**
      * Calculates the SBObj of the target.  This is needed to calculate the contrast of the target.
+     *
+     * @return ?float THe SBObj of the target
      */
-    public function calculateSBObj()
+    public function calculateSBObj(): ?float
     {
         if (($this->_magnitude) && ($this->_magnitude != 99.9) && (($this->_diam1 != 0) || ($this->_diam2 != 0))) {
             $SBObj = ($this->_magnitude + (2.5 * log10(2827.0 * ($this->_diam1 / 60) * ($this->_diam2 / 60))));
@@ -1155,5 +1263,144 @@ class Target
             $SBObj = null;
         }
         return $SBObj;
+    }
+
+    /**
+     * Calculates the contrast reserve of the target.
+     * If the contrast difference is < 0, the object is not visible.
+     *    contrast difference < -0.2 : Not visible
+     *     -0.2 < contrast diff < 0.1 : questionable
+     *     0.10 < contrast diff < 0.35 : Difficult
+     *     0.35 < contrast diff < 0.5 : Quite difficult to see
+     *     0.50 < contr diff < 1.0 : Easy to see
+     *     1.00 < contrast diff : Very easy to see
+     *
+     * @param float $SBObj         SBObj as calculated.
+     * @param float $sqm           The value from the Sky Quality Meter describing the sky darkness.
+     * @param float $diameter      The diameter of the used instrument (in mm)
+     * @param float $magnification The used magnification
+     */
+    public function calculateContrastReserve(float $SBObj, ?float $sqm, ?float $diameter, ?float $magnification): ?float
+    {
+        if (!$sqm) {
+            return null;
+        }
+        if (!$diameter) {
+            return null;
+        }
+
+        $aperIn = $diameter / 25.4;
+
+        // Minimum useful magnification
+        $SBB1 = $sqm - (5 * log10(2.833 * $aperIn));
+
+        $minObjArcmin = $this->_diam1 / 60.0;
+        $maxObjArcmin = $this->_diam2 / 60.0;
+
+        if ($minObjArcmin > $maxObjArcmin) {
+            $temp         = $minObjArcmin;
+            $minObjArcmin = $maxObjArcmin;
+            $maxObjArcmin = $temp;
+        }
+        $maxLog = 37;
+
+        // Log Object contrast
+        $logObjContrast = -0.4 * ($SBObj - $sqm);
+
+        // The preparations are finished, we can now start the calculations
+        $x     = $magnification;
+
+        $SBReduc    = 5 * log10($x);
+        $SBB        = $SBB1 + $SBReduc;
+
+        /* 2 dimensional interpolation of LTC array */
+        $ang    = $x * $minObjArcmin;
+        $logAng = log10($ang);
+        $SB     = $SBB;
+        $I      = 0;
+
+        /* int of surface brightness */
+        $intSB = (int) $SB;
+        /* surface brightness index A */
+        $SBIA = $intSB - 4;
+        /* min index must be at least 0 */
+        if ($SBIA < 0) {
+            $SBIA = 0;
+        }
+        /* max SBIA index cannot > 22 so that max SBIB <= 23 */
+        if ($SBIA > $this->_LTCSize - 2) {
+            $SBIA = $this->_LTCSize - 2;
+        }
+        /* surface brightness index B */
+        $SBIB = $SBIA + 1;
+
+        while ($I < $this->_angleSize && $logAng > $this->_angle[$I++]);
+
+        /* found 1st Angle[] value > LogAng, so back up 2 */
+        $I -= 2;
+        if ($I < 0) {
+            $I      = 0;
+            $logAng = $this->_angle[0];
+        }
+
+        /* ie, if LogAng = 4 and Angle[I] = 3 and Angle[I+1] = 5,
+        InterpAngle = .5, or .5 of the way between Angle[I] and Angle{I+1] */
+        $interpAngle = ($logAng - $this->_angle[$I])
+                    / ($this->_angle[$I + 1] - $this->_angle[$I]);
+        /* add 1 to I because first entry in LTC is
+        sky background brightness */
+        $interpA = $this->_LTC[$SBIA][$I + 1]
+                    + $interpAngle
+                    * ($this->_LTC[$SBIA][$I + 2]
+                    - $this->_LTC[$SBIA][$I + 1]);
+        $interpB = $this->_LTC[$SBIB][$I + 1]
+                    + $interpAngle
+                    * ($this->_LTC[$SBIB][$I + 2]
+                    - $this->_LTC[$SBIB][$I + 1]);
+        if ($SB < $this->_LTC[0][0]) {
+            $SB = $this->_LTC[0][0];
+        }
+        if ($intSB >= $this->_LTC[$this->_LTCSize - 1][0]) {
+            $logThreshContrast = $interpB
+                        + ($SB - $this->_LTC[$this->_LTCSize - 1][0])
+                        * ($interpB - $interpA);
+        } else {
+            $logThreshContrast = $interpA + ($SB - $intSB)
+                        * ($interpB - $interpA);
+        }
+
+        if ($logThreshContrast > $maxLog) {
+            $logThreshContrast = $maxLog;
+        } else {
+            if ($logThreshContrast < -$maxLog) {
+                $logThreshContrast = -$maxLog;
+            }
+        }
+
+        $logContrastDiff = $logObjContrast - $logThreshContrast;
+
+        return $logContrastDiff;
+    }
+
+    /**
+     * Calculates the detection magnification of the target.
+     *
+     * @param float $SBObj          SBObj as calculated.
+     * @param float $sqm            The value from the Sky Quality Meter describing the sky darkness.
+     * @param float $diameter       The diameter of the used instrument (in mm)
+     * @param array $magnifications An array with the possible magnifications
+     */
+    public function calculateBestMagnification(float $SBObj, ?float $sqm, ?float $diameter, array $magnifications): ?float
+    {
+        $bestContrast      = -999;
+        $bestMagnification = 0;
+        foreach ($magnifications as $magnification) {
+            $contrast = $this->calculateContrastReserve($SBObj, $sqm, $diameter, $magnification);
+            if ($contrast > $bestContrast) {
+                $bestContrast      = $contrast;
+                $bestMagnification = $magnification;
+            }
+        }
+        return $bestMagnification;
     }
 }