+++ tests/Unit/ContrastTest.php
@@ -96,4 +96,90 @@ class ContrastTest extends BaseTestCase
         $target->setMagnitude(9.2);
         $this->assertEqualsWithDelta(21.1182, $target->calculateSBObj(), 0.001);
     }
+
+    /**
+     * Test the calculation of the contrast.
+     *
+     * @return None
+     */
+    public function testCalculateContrast()
+    {
+        // Berk 59
+        $target = new Target();
+        $target->setDiameter(600);
+        $target->setMagnitude(11);
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(0.13, $target->calculateContrastReserve($target->calculateSBObj(), 22, 457, 118), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(-0.35, $target->calculateContrastReserve($target->calculateSBObj(), 20.15, 457, 473), 0.01);
+
+        // M 65
+        $target = new Target();
+        $target->setDiameter(540, 138);
+        $target->setMagnitude(9.2);
+        // SBObj = 21.1182
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(1.18, $target->calculateContrastReserve($target->calculateSBObj(), 22, 457, 66), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(0.70, $target->calculateContrastReserve($target->calculateSBObj(), 20.15, 457, 66), 0.01);
+
+        // M 82
+        $target = new Target();
+        $target->setDiameter(630, 306);
+        $target->setMagnitude(8.6);
+        // SBObj = 21.5502
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(1.20, $target->calculateContrastReserve($target->calculateSBObj(), 22, 457, 66), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(0.70, $target->calculateContrastReserve($target->calculateSBObj(), 20.15, 457, 66), 0.01);
+    }
+
+    /**
+     * Test the calculation of the best detection magnification.
+     *
+     * @return None
+     */
+    public function testCalculateBestMagnification()
+    {
+        $magnifications = [
+            66, 103, 158, 257, 411,
+            76, 118, 182, 296, 473,
+            133, 206, 317, 514, 823,
+        ];
+
+        // Berk 59
+        $target = new Target();
+        $target->setDiameter(600);
+        $target->setMagnitude(11);
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(133, $target->calculateBestMagnification($target->calculateSBObj(), 22, 457, $magnifications), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(473, $target->calculateBestMagnification($target->calculateSBObj(), 20.15, 457, $magnifications), 0.01);
+
+        // M 65
+        $target = new Target();
+        $target->setDiameter(540, 138);
+        $target->setMagnitude(9.2);
+        // SBObj = 21.1182
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(66, $target->calculateBestMagnification($target->calculateSBObj(), 22, 457, $magnifications), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(66, $target->calculateBestMagnification($target->calculateSBObj(), 20.15, 457, $magnifications), 0.01);
+
+        // M 82
+        $target = new Target();
+        $target->setDiameter(630, 306);
+        $target->setMagnitude(8.6);
+        // SBObj = 21.5502
+        // Diameter telescope: 457mm
+        // SQM / NELM location: 22 / 6.7
+        $this->assertEqualsWithDelta(66, $target->calculateBestMagnification($target->calculateSBObj(), 22, 457, $magnifications), 0.01);
+        // SQM / NELM location: 20.15 / 5.6
+        $this->assertEqualsWithDelta(66, $target->calculateBestMagnification($target->calculateSBObj(), 20.15, 457, $magnifications), 0.01);
+    }
 }